+++ code/react/src/util/index.ts
@@ -1,1 +1,2 @@
-export * from './util';
+export * from './util.copyToClipboard';
+export * from './util.containsFocus';
